feat: Add FIA configuration files for Proxmox host
This commit adds the FIA configuration files for the Proxmox host. It
includes the README.md file documenting the purpose of FIA, ferrari.tf
file defining the resources for the Ferrari container, and providers.tf
file specifying the Proxmox provider and its configuration.

The addition of these files enables seamless management and deployment
of virtual machines and Linux containers on the Proxmox instance using
Terraform and the Proxmox provider.